Move code from module to service.
